--- .idea/.name
@@ -1,1 +1,0 @@
-Fall2022-Group20
