+++ README.md
@@ -1,4 +1,3 @@
-# Fall2022-Group20
 # Growth+
 
 ## About Growth +
@@ -14,7 +13,6 @@ Students in Africa have fewer opportunities to gain these skills and as a result
 
 ### UI Layer
 * Victor Yue
-* Zack Orooji
 
 ### Data/Domain Layer
 * Julio Alva
