+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Fall2022Group20"
+        android:theme="@style/Theme.GrowthPlus"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
