+++ app/src/main/java/com/GrowthPlus/dataAccessLayer/Language/LanguageSchema.java
@@ -1,0 +1,1691 @@
+package com.GrowthPlus.dataAccessLayer.Language;
+
+import io.realm.RealmObject;
+import io.realm.annotations.PrimaryKey;
+
+public class LanguageSchema extends RealmObject {
+    @PrimaryKey
+    String languageId;
+    String languageName;
+    String	growthPlus;
+    String	parent;
+    String	child;
+    String	points;
+    String	pin;
+    String	create;
+    String	confirm;
+    String	numbers;
+    String	counting;
+    String	addition;
+    String	subtraction;
+    String	multiplication;
+    String	division;
+    String	measurements;
+    String	weights;
+    String	money;
+    String	time;
+    String	shapes;
+    String	angles;
+    String	name;
+    String	level ;
+    String	lesson;
+    String	game;
+    String	roadmap;
+    String	francs;
+    String	cfa;
+    String	centimeter ;
+    String	decimeter;
+    String	meter;
+    String	decameter;
+    String	hectometer	;
+    String	centigram	;
+    String	decigram;
+    String	gram;
+    String	decagram;
+    String	hectogram;
+    String	centiliter;
+    String	deciliter;
+    String	liter;
+    String	millimeter;
+    String	kilometer;
+    String	milligram;
+    String	kilogram ;
+    String	milliliter;
+    String	kiloliter;
+    String	unit;
+    String	unitOfTen 	;
+    String	unitOfOneHundred	;
+    String	clock;
+    String	hour;
+    String	minute;
+    String	day	;
+    String	circle;
+    String	ray	;
+    String	diameter;
+    String	center;
+    String	rectangle;
+    String	width;
+    String	length;
+    String	side;
+    String	square;
+    String	triangle;
+    String	perimeter;
+    String	parallelLines;
+    String	perpendicularLines;
+    String	rightAngle;
+    String	area;
+    String	one;
+    String	two;
+    String	three;
+    String	four;
+    String	five;
+    String	six;
+    String	seven;
+    String	eight;
+    String	nine;
+    String	ten;
+    String	eleven;
+    String	twelve;
+    String	thirteen;
+    String	fourteen;
+    String	fifteen	;
+    String	sixteen	;
+    String	seventeen ;
+    String	eighteen;
+    String	nineteen;
+    String	twenty;
+    String	twentyOne	;
+    String	twentyTwo	;
+    String	twentyThree	;
+    String	twentyFour	;
+    String	twentyFive	;
+    String	twentySix	;
+    String	twentySeven	;
+    String	twentyEight	;
+    String	twentyNine	;
+    String	thirty	;
+    String	thirtyOne 	;
+    String	thirtyTwo 	;
+    String	thirtyThree ;
+    String	thirtyFour 	;
+    String	thirtyFive 	;
+    String	thirtySix 	;
+    String	thirtySeven ;
+    String	thirtyEight ;
+    String	thirtyNine 	;
+    String	forty	;
+    String	fortyOne	;
+    String	fortyTwo	;
+    String	fortyThree	;
+    String	fortyFour	;
+    String	fortyFive	;
+    String	fortySix	;
+    String	fortySeven	;
+    String	fortyEight	;
+    String	fortyNine	;
+    String	fifty 	;
+    String	fiftyOne;
+    String	fiftyTwo;
+    String	fiftyThree	;
+    String	fiftyFour	;
+    String	fiftyFive	;
+    String	fiftySix	;
+    String	fiftySeven	;
+    String	fiftyEight	;
+    String	fiftyNine	;
+    String	sixty	;
+    String	sixtyOne;
+    String	sixtyTwo	;
+    String	sixtyThree	;
+    String	sixtyFour	;
+    String	sixtyFive	;
+    String	sixtySix;
+    String	sixtySeven	;
+    String	sixtyEight	;
+    String	sixtyNine	;
+    String	seventy	;
+    String	seventyOne	;
+    String	seventyTwo	;
+    String	seventyThree;
+    String	seventyFour	;
+    String	seventyFive	;
+    String	seventySix	;
+    String	seventySeven;
+    String	seventyEight;
+    String	seventyNine	;
+    String	eighty 	;
+    String	eightyOne	;
+    String	eightyTwo	;
+    String	eightyThree	;
+    String	eightyFour	;
+    String	eightyFive	;
+    String	eightySix	;
+    String	eightySeven	;
+    String	eightyEight	;
+    String	eightyNine	;
+    String	ninety ;
+    String	ninetyOne	;
+    String	ninetyTwo	;
+    String	ninetyThree	;
+    String	ninetyFour	;
+    String	ninetyFive	;
+    String	ninetySix	;
+    String	ninetySeven	;
+    String	ninetyEight	;
+    String	ninetyNine	;
+    String	oneHundred	;
+    String	twoHundred	;
+    String	threeHundred;
+    String	fourHundred	;
+    String	fiveHundred	;
+    String	sixHundred	;
+    String	sevenHundred;
+    String	eightHundred;
+    String	nineHundred	;
+    String	oneThousand	;
+    String	twoThousand	;
+    String	threeThousand;
+    String	fourThousand;
+    String	fiveThousand;
+    String	sixThousand	;
+    String	sevenThousand;
+    String	eightThousand;
+    String	nineThousand;
+    String	tenThousand	;
+
+    public String getLanguageId() {
+        return languageId;
+    }
+
+    public void setLanguageId(String languageId) {
+        this.languageId = languageId;
+    }
+
+    public String getLanguageName() {
+        return languageName;
+    }
+
+    public void setLanguageName(String languageName) {
+        this.languageName = languageName;
+    }
+
+    public String getGrowthPlus() {
+        return growthPlus;
+    }
+
+    public void setGrowthPlus(String growthPlus) {
+        this.growthPlus = growthPlus;
+    }
+
+    public String getParent() {
+        return parent;
+    }
+
+    public void setParent(String parent) {
+        this.parent = parent;
+    }
+
+    public String getChild() {
+        return child;
+    }
+
+    public void setChild(String child) {
+        this.child = child;
+    }
+
+    public String getPoints() {
+        return points;
+    }
+
+    public void setPoints(String points) {
+        this.points = points;
+    }
+
+    public String getPin() {
+        return pin;
+    }
+
+    public void setPin(String pin) {
+        this.pin = pin;
+    }
+
+    public String getCreate() {
+        return create;
+    }
+
+    public void setCreate(String create) {
+        this.create = create;
+    }
+
+    public String getConfirm() {
+        return confirm;
+    }
+
+    public void setConfirm(String confirm) {
+        this.confirm = confirm;
+    }
+
+    public String getNumbers() {
+        return numbers;
+    }
+
+    public void setNumbers(String numbers) {
+        this.numbers = numbers;
+    }
+
+    public String getCounting() {
+        return counting;
+    }
+
+    public void setCounting(String counting) {
+        this.counting = counting;
+    }
+
+    public String getAddition() {
+        return addition;
+    }
+
+    public void setAddition(String addition) {
+        this.addition = addition;
+    }
+
+    public String getSubtraction() {
+        return subtraction;
+    }
+
+    public void setSubtraction(String subtraction) {
+        this.subtraction = subtraction;
+    }
+
+    public String getMultiplication() {
+        return multiplication;
+    }
+
+    public void setMultiplication(String multiplication) {
+        this.multiplication = multiplication;
+    }
+
+    public String getDivision() {
+        return division;
+    }
+
+    public void setDivision(String division) {
+        this.division = division;
+    }
+
+    public String getMeasurements() {
+        return measurements;
+    }
+
+    public void setMeasurements(String measurements) {
+        this.measurements = measurements;
+    }
+
+    public String getWeights() {
+        return weights;
+    }
+
+    public void setWeights(String weights) {
+        this.weights = weights;
+    }
+
+    public String getMoney() {
+        return money;
+    }
+
+    public void setMoney(String money) {
+        this.money = money;
+    }
+
+    public String getTime() {
+        return time;
+    }
+
+    public void setTime(String time) {
+        this.time = time;
+    }
+
+    public String getShapes() {
+        return shapes;
+    }
+
+    public void setShapes(String shapes) {
+        this.shapes = shapes;
+    }
+
+    public String getAngles() {
+        return angles;
+    }
+
+    public void setAngles(String angles) {
+        this.angles = angles;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getLevel() {
+        return level;
+    }
+
+    public void setLevel(String level) {
+        this.level = level;
+    }
+
+    public String getLesson() {
+        return lesson;
+    }
+
+    public void setLesson(String lesson) {
+        this.lesson = lesson;
+    }
+
+    public String getGame() {
+        return game;
+    }
+
+    public void setGame(String game) {
+        this.game = game;
+    }
+
+    public String getRoadmap() {
+        return roadmap;
+    }
+
+    public void setRoadmap(String roadmap) {
+        this.roadmap = roadmap;
+    }
+
+    public String getFrancs() {
+        return francs;
+    }
+
+    public void setFrancs(String francs) {
+        this.francs = francs;
+    }
+
+    public String getCfa() {
+        return cfa;
+    }
+
+    public void setCfa(String cfa) {
+        this.cfa = cfa;
+    }
+
+    public String getCentimeter() {
+        return centimeter;
+    }
+
+    public void setCentimeter(String centimeter) {
+        this.centimeter = centimeter;
+    }
+
+    public String getDecimeter() {
+        return decimeter;
+    }
+
+    public void setDecimeter(String decimeter) {
+        this.decimeter = decimeter;
+    }
+
+    public String getMeter() {
+        return meter;
+    }
+
+    public void setMeter(String meter) {
+        this.meter = meter;
+    }
+
+    public String getDecameter() {
+        return decameter;
+    }
+
+    public void setDecameter(String decameter) {
+        this.decameter = decameter;
+    }
+
+    public String getHectometer() {
+        return hectometer;
+    }
+
+    public void setHectometer(String hectometer) {
+        this.hectometer = hectometer;
+    }
+
+    public String getCentigram() {
+        return centigram;
+    }
+
+    public void setCentigram(String centigram) {
+        this.centigram = centigram;
+    }
+
+    public String getDecigram() {
+        return decigram;
+    }
+
+    public void setDecigram(String decigram) {
+        this.decigram = decigram;
+    }
+
+    public String getGram() {
+        return gram;
+    }
+
+    public void setGram(String gram) {
+        this.gram = gram;
+    }
+
+    public String getDecagram() {
+        return decagram;
+    }
+
+    public void setDecagram(String decagram) {
+        this.decagram = decagram;
+    }
+
+    public String getHectogram() {
+        return hectogram;
+    }
+
+    public void setHectogram(String hectogram) {
+        this.hectogram = hectogram;
+    }
+
+    public String getCentiliter() {
+        return centiliter;
+    }
+
+    public void setCentiliter(String centiliter) {
+        this.centiliter = centiliter;
+    }
+
+    public String getDeciliter() {
+        return deciliter;
+    }
+
+    public void setDeciliter(String deciliter) {
+        this.deciliter = deciliter;
+    }
+
+    public String getLiter() {
+        return liter;
+    }
+
+    public void setLiter(String liter) {
+        this.liter = liter;
+    }
+
+    public String getMillimeter() {
+        return millimeter;
+    }
+
+    public void setMillimeter(String millimeter) {
+        this.millimeter = millimeter;
+    }
+
+    public String getKilometer() {
+        return kilometer;
+    }
+
+    public void setKilometer(String kilometer) {
+        this.kilometer = kilometer;
+    }
+
+    public String getMilligram() {
+        return milligram;
+    }
+
+    public void setMilligram(String milligram) {
+        this.milligram = milligram;
+    }
+
+    public String getKilogram() {
+        return kilogram;
+    }
+
+    public void setKilogram(String kilogram) {
+        this.kilogram = kilogram;
+    }
+
+    public String getMilliliter() {
+        return milliliter;
+    }
+
+    public void setMilliliter(String milliliter) {
+        this.milliliter = milliliter;
+    }
+
+    public String getKiloliter() {
+        return kiloliter;
+    }
+
+    public void setKiloliter(String kiloliter) {
+        this.kiloliter = kiloliter;
+    }
+
+    public String getUnit() {
+        return unit;
+    }
+
+    public void setUnit(String unit) {
+        this.unit = unit;
+    }
+
+    public String getUnitOfTen() {
+        return unitOfTen;
+    }
+
+    public void setUnitOfTen(String unitOfTen) {
+        this.unitOfTen = unitOfTen;
+    }
+
+    public String getUnitOfOneHundred() {
+        return unitOfOneHundred;
+    }
+
+    public void setUnitOfOneHundred(String unitOfOneHundred) {
+        this.unitOfOneHundred = unitOfOneHundred;
+    }
+
+    public String getClock() {
+        return clock;
+    }
+
+    public void setClock(String clock) {
+        this.clock = clock;
+    }
+
+    public String getHour() {
+        return hour;
+    }
+
+    public void setHour(String hour) {
+        this.hour = hour;
+    }
+
+    public String getMinute() {
+        return minute;
+    }
+
+    public void setMinute(String minute) {
+        this.minute = minute;
+    }
+
+    public String getDay() {
+        return day;
+    }
+
+    public void setDay(String day) {
+        this.day = day;
+    }
+
+    public String getCircle() {
+        return circle;
+    }
+
+    public void setCircle(String circle) {
+        this.circle = circle;
+    }
+
+    public String getRay() {
+        return ray;
+    }
+
+    public void setRay(String ray) {
+        this.ray = ray;
+    }
+
+    public String getDiameter() {
+        return diameter;
+    }
+
+    public void setDiameter(String diameter) {
+        this.diameter = diameter;
+    }
+
+    public String getCenter() {
+        return center;
+    }
+
+    public void setCenter(String center) {
+        this.center = center;
+    }
+
+    public String getRectangle() {
+        return rectangle;
+    }
+
+    public void setRectangle(String rectangle) {
+        this.rectangle = rectangle;
+    }
+
+    public String getWidth() {
+        return width;
+    }
+
+    public void setWidth(String width) {
+        this.width = width;
+    }
+
+    public String getLength() {
+        return length;
+    }
+
+    public void setLength(String length) {
+        this.length = length;
+    }
+
+    public String getSide() {
+        return side;
+    }
+
+    public void setSide(String side) {
+        this.side = side;
+    }
+
+    public String getSquare() {
+        return square;
+    }
+
+    public void setSquare(String square) {
+        this.square = square;
+    }
+
+    public String getTriangle() {
+        return triangle;
+    }
+
+    public void setTriangle(String triangle) {
+        this.triangle = triangle;
+    }
+
+    public String getPerimeter() {
+        return perimeter;
+    }
+
+    public void setPerimeter(String perimeter) {
+        this.perimeter = perimeter;
+    }
+
+    public String getParallelLines() {
+        return parallelLines;
+    }
+
+    public void setParallelLines(String parallelLines) {
+        this.parallelLines = parallelLines;
+    }
+
+    public String getPerpendicularLines() {
+        return perpendicularLines;
+    }
+
+    public void setPerpendicularLines(String perpendicularLines) {
+        this.perpendicularLines = perpendicularLines;
+    }
+
+    public String getRightAngle() {
+        return rightAngle;
+    }
+
+    public void setRightAngle(String rightAngle) {
+        this.rightAngle = rightAngle;
+    }
+
+    public String getArea() {
+        return area;
+    }
+
+    public void setArea(String area) {
+        this.area = area;
+    }
+
+    public String getOne() {
+        return one;
+    }
+
+    public void setOne(String one) {
+        this.one = one;
+    }
+
+    public String getTwo() {
+        return two;
+    }
+
+    public void setTwo(String two) {
+        this.two = two;
+    }
+
+    public String getThree() {
+        return three;
+    }
+
+    public void setThree(String three) {
+        this.three = three;
+    }
+
+    public String getFour() {
+        return four;
+    }
+
+    public void setFour(String four) {
+        this.four = four;
+    }
+
+    public String getFive() {
+        return five;
+    }
+
+    public void setFive(String five) {
+        this.five = five;
+    }
+
+    public String getSix() {
+        return six;
+    }
+
+    public void setSix(String six) {
+        this.six = six;
+    }
+
+    public String getSeven() {
+        return seven;
+    }
+
+    public void setSeven(String seven) {
+        this.seven = seven;
+    }
+
+    public String getEight() {
+        return eight;
+    }
+
+    public void setEight(String eight) {
+        this.eight = eight;
+    }
+
+    public String getNine() {
+        return nine;
+    }
+
+    public void setNine(String nine) {
+        this.nine = nine;
+    }
+
+    public String getTen() {
+        return ten;
+    }
+
+    public void setTen(String ten) {
+        this.ten = ten;
+    }
+
+    public String getEleven() {
+        return eleven;
+    }
+
+    public void setEleven(String eleven) {
+        this.eleven = eleven;
+    }
+
+    public String getTwelve() {
+        return twelve;
+    }
+
+    public void setTwelve(String twelve) {
+        this.twelve = twelve;
+    }
+
+    public String getThirteen() {
+        return thirteen;
+    }
+
+    public void setThirteen(String thirteen) {
+        this.thirteen = thirteen;
+    }
+
+    public String getFourteen() {
+        return fourteen;
+    }
+
+    public void setFourteen(String fourteen) {
+        this.fourteen = fourteen;
+    }
+
+    public String getFifteen() {
+        return fifteen;
+    }
+
+    public void setFifteen(String fifteen) {
+        this.fifteen = fifteen;
+    }
+
+    public String getSixteen() {
+        return sixteen;
+    }
+
+    public void setSixteen(String sixteen) {
+        this.sixteen = sixteen;
+    }
+
+    public String getSeventeen() {
+        return seventeen;
+    }
+
+    public void setSeventeen(String seventeen) {
+        this.seventeen = seventeen;
+    }
+
+    public String getEighteen() {
+        return eighteen;
+    }
+
+    public void setEighteen(String eighteen) {
+        this.eighteen = eighteen;
+    }
+
+    public String getNineteen() {
+        return nineteen;
+    }
+
+    public void setNineteen(String nineteen) {
+        this.nineteen = nineteen;
+    }
+
+    public String getTwenty() {
+        return twenty;
+    }
+
+    public void setTwenty(String twenty) {
+        this.twenty = twenty;
+    }
+
+    public String getTwentyOne() {
+        return twentyOne;
+    }
+
+    public void setTwentyOne(String twentyOne) {
+        this.twentyOne = twentyOne;
+    }
+
+    public String getTwentyTwo() {
+        return twentyTwo;
+    }
+
+    public void setTwentyTwo(String twentyTwo) {
+        this.twentyTwo = twentyTwo;
+    }
+
+    public String getTwentyThree() {
+        return twentyThree;
+    }
+
+    public void setTwentyThree(String twentyThree) {
+        this.twentyThree = twentyThree;
+    }
+
+    public String getTwentyFour() {
+        return twentyFour;
+    }
+
+    public void setTwentyFour(String twentyFour) {
+        this.twentyFour = twentyFour;
+    }
+
+    public String getTwentyFive() {
+        return twentyFive;
+    }
+
+    public void setTwentyFive(String twentyFive) {
+        this.twentyFive = twentyFive;
+    }
+
+    public String getTwentySix() {
+        return twentySix;
+    }
+
+    public void setTwentySix(String twentySix) {
+        this.twentySix = twentySix;
+    }
+
+    public String getTwentySeven() {
+        return twentySeven;
+    }
+
+    public void setTwentySeven(String twentySeven) {
+        this.twentySeven = twentySeven;
+    }
+
+    public String getTwentyEight() {
+        return twentyEight;
+    }
+
+    public void setTwentyEight(String twentyEight) {
+        this.twentyEight = twentyEight;
+    }
+
+    public String getTwentyNine() {
+        return twentyNine;
+    }
+
+    public void setTwentyNine(String twentyNine) {
+        this.twentyNine = twentyNine;
+    }
+
+    public String getThirty() {
+        return thirty;
+    }
+
+    public void setThirty(String thirty) {
+        this.thirty = thirty;
+    }
+
+    public String getThirtyOne() {
+        return thirtyOne;
+    }
+
+    public void setThirtyOne(String thirtyOne) {
+        this.thirtyOne = thirtyOne;
+    }
+
+    public String getThirtyTwo() {
+        return thirtyTwo;
+    }
+
+    public void setThirtyTwo(String thirtyTwo) {
+        this.thirtyTwo = thirtyTwo;
+    }
+
+    public String getThirtyThree() {
+        return thirtyThree;
+    }
+
+    public void setThirtyThree(String thirtyThree) {
+        this.thirtyThree = thirtyThree;
+    }
+
+    public String getThirtyFour() {
+        return thirtyFour;
+    }
+
+    public void setThirtyFour(String thirtyFour) {
+        this.thirtyFour = thirtyFour;
+    }
+
+    public String getThirtyFive() {
+        return thirtyFive;
+    }
+
+    public void setThirtyFive(String thirtyFive) {
+        this.thirtyFive = thirtyFive;
+    }
+
+    public String getThirtySix() {
+        return thirtySix;
+    }
+
+    public void setThirtySix(String thirtySix) {
+        this.thirtySix = thirtySix;
+    }
+
+    public String getThirtySeven() {
+        return thirtySeven;
+    }
+
+    public void setThirtySeven(String thirtySeven) {
+        this.thirtySeven = thirtySeven;
+    }
+
+    public String getThirtyEight() {
+        return thirtyEight;
+    }
+
+    public void setThirtyEight(String thirtyEight) {
+        this.thirtyEight = thirtyEight;
+    }
+
+    public String getThirtyNine() {
+        return thirtyNine;
+    }
+
+    public void setThirtyNine(String thirtyNine) {
+        this.thirtyNine = thirtyNine;
+    }
+
+    public String getForty() {
+        return forty;
+    }
+
+    public void setForty(String forty) {
+        this.forty = forty;
+    }
+
+    public String getFortyOne() {
+        return fortyOne;
+    }
+
+    public void setFortyOne(String fortyOne) {
+        this.fortyOne = fortyOne;
+    }
+
+    public String getFortyTwo() {
+        return fortyTwo;
+    }
+
+    public void setFortyTwo(String fortyTwo) {
+        this.fortyTwo = fortyTwo;
+    }
+
+    public String getFortyThree() {
+        return fortyThree;
+    }
+
+    public void setFortyThree(String fortyThree) {
+        this.fortyThree = fortyThree;
+    }
+
+    public String getFortyFour() {
+        return fortyFour;
+    }
+
+    public void setFortyFour(String fortyFour) {
+        this.fortyFour = fortyFour;
+    }
+
+    public String getFortyFive() {
+        return fortyFive;
+    }
+
+    public void setFortyFive(String fortyFive) {
+        this.fortyFive = fortyFive;
+    }
+
+    public String getFortySix() {
+        return fortySix;
+    }
+
+    public void setFortySix(String fortySix) {
+        this.fortySix = fortySix;
+    }
+
+    public String getFortySeven() {
+        return fortySeven;
+    }
+
+    public void setFortySeven(String fortySeven) {
+        this.fortySeven = fortySeven;
+    }
+
+    public String getFortyEight() {
+        return fortyEight;
+    }
+
+    public void setFortyEight(String fortyEight) {
+        this.fortyEight = fortyEight;
+    }
+
+    public String getFortyNine() {
+        return fortyNine;
+    }
+
+    public void setFortyNine(String fortyNine) {
+        this.fortyNine = fortyNine;
+    }
+
+    public String getFifty() {
+        return fifty;
+    }
+
+    public void setFifty(String fifty) {
+        this.fifty = fifty;
+    }
+
+    public String getFiftyOne() {
+        return fiftyOne;
+    }
+
+    public void setFiftyOne(String fiftyOne) {
+        this.fiftyOne = fiftyOne;
+    }
+
+    public String getFiftyTwo() {
+        return fiftyTwo;
+    }
+
+    public void setFiftyTwo(String fiftyTwo) {
+        this.fiftyTwo = fiftyTwo;
+    }
+
+    public String getFiftyThree() {
+        return fiftyThree;
+    }
+
+    public void setFiftyThree(String fiftyThree) {
+        this.fiftyThree = fiftyThree;
+    }
+
+    public String getFiftyFour() {
+        return fiftyFour;
+    }
+
+    public void setFiftyFour(String fiftyFour) {
+        this.fiftyFour = fiftyFour;
+    }
+
+    public String getFiftyFive() {
+        return fiftyFive;
+    }
+
+    public void setFiftyFive(String fiftyFive) {
+        this.fiftyFive = fiftyFive;
+    }
+
+    public String getFiftySix() {
+        return fiftySix;
+    }
+
+    public void setFiftySix(String fiftySix) {
+        this.fiftySix = fiftySix;
+    }
+
+    public String getFiftySeven() {
+        return fiftySeven;
+    }
+
+    public void setFiftySeven(String fiftySeven) {
+        this.fiftySeven = fiftySeven;
+    }
+
+    public String getFiftyEight() {
+        return fiftyEight;
+    }
+
+    public void setFiftyEight(String fiftyEight) {
+        this.fiftyEight = fiftyEight;
+    }
+
+    public String getFiftyNine() {
+        return fiftyNine;
+    }
+
+    public void setFiftyNine(String fiftyNine) {
+        this.fiftyNine = fiftyNine;
+    }
+
+    public String getSixty() {
+        return sixty;
+    }
+
+    public void setSixty(String sixty) {
+        this.sixty = sixty;
+    }
+
+    public String getSixtyOne() {
+        return sixtyOne;
+    }
+
+    public void setSixtyOne(String sixtyOne) {
+        this.sixtyOne = sixtyOne;
+    }
+
+    public String getSixtyTwo() {
+        return sixtyTwo;
+    }
+
+    public void setSixtyTwo(String sixtyTwo) {
+        this.sixtyTwo = sixtyTwo;
+    }
+
+    public String getSixtyThree() {
+        return sixtyThree;
+    }
+
+    public void setSixtyThree(String sixtyThree) {
+        this.sixtyThree = sixtyThree;
+    }
+
+    public String getSixtyFour() {
+        return sixtyFour;
+    }
+
+    public void setSixtyFour(String sixtyFour) {
+        this.sixtyFour = sixtyFour;
+    }
+
+    public String getSixtyFive() {
+        return sixtyFive;
+    }
+
+    public void setSixtyFive(String sixtyFive) {
+        this.sixtyFive = sixtyFive;
+    }
+
+    public String getSixtySix() {
+        return sixtySix;
+    }
+
+    public void setSixtySix(String sixtySix) {
+        this.sixtySix = sixtySix;
+    }
+
+    public String getSixtySeven() {
+        return sixtySeven;
+    }
+
+    public void setSixtySeven(String sixtySeven) {
+        this.sixtySeven = sixtySeven;
+    }
+
+    public String getSixtyEight() {
+        return sixtyEight;
+    }
+
+    public void setSixtyEight(String sixtyEight) {
+        this.sixtyEight = sixtyEight;
+    }
+
+    public String getSixtyNine() {
+        return sixtyNine;
+    }
+
+    public void setSixtyNine(String sixtyNine) {
+        this.sixtyNine = sixtyNine;
+    }
+
+    public String getSeventy() {
+        return seventy;
+    }
+
+    public void setSeventy(String seventy) {
+        this.seventy = seventy;
+    }
+
+    public String getSeventyOne() {
+        return seventyOne;
+    }
+
+    public void setSeventyOne(String seventyOne) {
+        this.seventyOne = seventyOne;
+    }
+
+    public String getSeventyTwo() {
+        return seventyTwo;
+    }
+
+    public void setSeventyTwo(String seventyTwo) {
+        this.seventyTwo = seventyTwo;
+    }
+
+    public String getSeventyThree() {
+        return seventyThree;
+    }
+
+    public void setSeventyThree(String seventyThree) {
+        this.seventyThree = seventyThree;
+    }
+
+    public String getSeventyFour() {
+        return seventyFour;
+    }
+
+    public void setSeventyFour(String seventyFour) {
+        this.seventyFour = seventyFour;
+    }
+
+    public String getSeventyFive() {
+        return seventyFive;
+    }
+
+    public void setSeventyFive(String seventyFive) {
+        this.seventyFive = seventyFive;
+    }
+
+    public String getSeventySix() {
+        return seventySix;
+    }
+
+    public void setSeventySix(String seventySix) {
+        this.seventySix = seventySix;
+    }
+
+    public String getSeventySeven() {
+        return seventySeven;
+    }
+
+    public void setSeventySeven(String seventySeven) {
+        this.seventySeven = seventySeven;
+    }
+
+    public String getSeventyEight() {
+        return seventyEight;
+    }
+
+    public void setSeventyEight(String seventyEight) {
+        this.seventyEight = seventyEight;
+    }
+
+    public String getSeventyNine() {
+        return seventyNine;
+    }
+
+    public void setSeventyNine(String seventyNine) {
+        this.seventyNine = seventyNine;
+    }
+
+    public String getEighty() {
+        return eighty;
+    }
+
+    public void setEighty(String eighty) {
+        this.eighty = eighty;
+    }
+
+    public String getEightyOne() {
+        return eightyOne;
+    }
+
+    public void setEightyOne(String eightyOne) {
+        this.eightyOne = eightyOne;
+    }
+
+    public String getEightyTwo() {
+        return eightyTwo;
+    }
+
+    public void setEightyTwo(String eightyTwo) {
+        this.eightyTwo = eightyTwo;
+    }
+
+    public String getEightyThree() {
+        return eightyThree;
+    }
+
+    public void setEightyThree(String eightyThree) {
+        this.eightyThree = eightyThree;
+    }
+
+    public String getEightyFour() {
+        return eightyFour;
+    }
+
+    public void setEightyFour(String eightyFour) {
+        this.eightyFour = eightyFour;
+    }
+
+    public String getEightyFive() {
+        return eightyFive;
+    }
+
+    public void setEightyFive(String eightyFive) {
+        this.eightyFive = eightyFive;
+    }
+
+    public String getEightySix() {
+        return eightySix;
+    }
+
+    public void setEightySix(String eightySix) {
+        this.eightySix = eightySix;
+    }
+
+    public String getEightySeven() {
+        return eightySeven;
+    }
+
+    public void setEightySeven(String eightySeven) {
+        this.eightySeven = eightySeven;
+    }
+
+    public String getEightyEight() {
+        return eightyEight;
+    }
+
+    public void setEightyEight(String eightyEight) {
+        this.eightyEight = eightyEight;
+    }
+
+    public String getEightyNine() {
+        return eightyNine;
+    }
+
+    public void setEightyNine(String eightyNine) {
+        this.eightyNine = eightyNine;
+    }
+
+    public String getNinety() {
+        return ninety;
+    }
+
+    public void setNinety(String ninety) {
+        this.ninety = ninety;
+    }
+
+    public String getNinetyOne() {
+        return ninetyOne;
+    }
+
+    public void setNinetyOne(String ninetyOne) {
+        this.ninetyOne = ninetyOne;
+    }
+
+    public String getNinetyTwo() {
+        return ninetyTwo;
+    }
+
+    public void setNinetyTwo(String ninetyTwo) {
+        this.ninetyTwo = ninetyTwo;
+    }
+
+    public String getNinetyThree() {
+        return ninetyThree;
+    }
+
+    public void setNinetyThree(String ninetyThree) {
+        this.ninetyThree = ninetyThree;
+    }
+
+    public String getNinetyFour() {
+        return ninetyFour;
+    }
+
+    public void setNinetyFour(String ninetyFour) {
+        this.ninetyFour = ninetyFour;
+    }
+
+    public String getNinetyFive() {
+        return ninetyFive;
+    }
+
+    public void setNinetyFive(String ninetyFive) {
+        this.ninetyFive = ninetyFive;
+    }
+
+    public String getNinetySix() {
+        return ninetySix;
+    }
+
+    public void setNinetySix(String ninetySix) {
+        this.ninetySix = ninetySix;
+    }
+
+    public String getNinetySeven() {
+        return ninetySeven;
+    }
+
+    public void setNinetySeven(String ninetySeven) {
+        this.ninetySeven = ninetySeven;
+    }
+
+    public String getNinetyEight() {
+        return ninetyEight;
+    }
+
+    public void setNinetyEight(String ninetyEight) {
+        this.ninetyEight = ninetyEight;
+    }
+
+    public String getNinetyNine() {
+        return ninetyNine;
+    }
+
+    public void setNinetyNine(String ninetyNine) {
+        this.ninetyNine = ninetyNine;
+    }
+
+    public String getOneHundred() {
+        return oneHundred;
+    }
+
+    public void setOneHundred(String oneHundred) {
+        this.oneHundred = oneHundred;
+    }
+
+    public String getTwoHundred() {
+        return twoHundred;
+    }
+
+    public void setTwoHundred(String twoHundred) {
+        this.twoHundred = twoHundred;
+    }
+
+    public String getThreeHundred() {
+        return threeHundred;
+    }
+
+    public void setThreeHundred(String threeHundred) {
+        this.threeHundred = threeHundred;
+    }
+
+    public String getFourHundred() {
+        return fourHundred;
+    }
+
+    public void setFourHundred(String fourHundred) {
+        this.fourHundred = fourHundred;
+    }
+
+    public String getFiveHundred() {
+        return fiveHundred;
+    }
+
+    public void setFiveHundred(String fiveHundred) {
+        this.fiveHundred = fiveHundred;
+    }
+
+    public String getSixHundred() {
+        return sixHundred;
+    }
+
+    public void setSixHundred(String sixHundred) {
+        this.sixHundred = sixHundred;
+    }
+
+    public String getSevenHundred() {
+        return sevenHundred;
+    }
+
+    public void setSevenHundred(String sevenHundred) {
+        this.sevenHundred = sevenHundred;
+    }
+
+    public String getEightHundred() {
+        return eightHundred;
+    }
+
+    public void setEightHundred(String eightHundred) {
+        this.eightHundred = eightHundred;
+    }
+
+    public String getNineHundred() {
+        return nineHundred;
+    }
+
+    public void setNineHundred(String nineHundred) {
+        this.nineHundred = nineHundred;
+    }
+
+    public String getOneThousand() {
+        return oneThousand;
+    }
+
+    public void setOneThousand(String oneThousand) {
+        this.oneThousand = oneThousand;
+    }
+
+    public String getTwoThousand() {
+        return twoThousand;
+    }
+
+    public void setTwoThousand(String twoThousand) {
+        this.twoThousand = twoThousand;
+    }
+
+    public String getThreeThousand() {
+        return threeThousand;
+    }
+
+    public void setThreeThousand(String threeThousand) {
+        this.threeThousand = threeThousand;
+    }
+
+    public String getFourThousand() {
+        return fourThousand;
+    }
+
+    public void setFourThousand(String fourThousand) {
+        this.fourThousand = fourThousand;
+    }
+
+    public String getFiveThousand() {
+        return fiveThousand;
+    }
+
+    public void setFiveThousand(String fiveThousand) {
+        this.fiveThousand = fiveThousand;
+    }
+
+    public String getSixThousand() {
+        return sixThousand;
+    }
+
+    public void setSixThousand(String sixThousand) {
+        this.sixThousand = sixThousand;
+    }
+
+    public String getSevenThousand() {
+        return sevenThousand;
+    }
+
+    public void setSevenThousand(String sevenThousand) {
+        this.sevenThousand = sevenThousand;
+    }
+
+    public String getEightThousand() {
+        return eightThousand;
+    }
+
+    public void setEightThousand(String eightThousand) {
+        this.eightThousand = eightThousand;
+    }
+
+    public String getNineThousand() {
+        return nineThousand;
+    }
+
+    public void setNineThousand(String nineThousand) {
+        this.nineThousand = nineThousand;
+    }
+
+    public String getTenThousand() {
+        return tenThousand;
+    }
+
+    public void setTenThousand(String tenThousand) {
+        this.tenThousand = tenThousand;
+    }
+}
