+++ app/src/main/java/com/GrowthPlus/dataAccessLayer/Report/ReportSchema.java
@@ -1,0 +1,45 @@
+package com.GrowthPlus.dataAccessLayer.Report;
+
+import io.realm.RealmObject;
+import io.realm.annotations.PrimaryKey;
+
+public class ReportSchema extends RealmObject {
+
+    @PrimaryKey
+    private String reportId;
+    private String childName;
+    private Integer childScore;
+    private String childId;
+
+    public String getReportId() {
+        return reportId;
+    }
+
+    public void setReportId(String reportId) {
+        this.reportId = reportId;
+    }
+
+    public String getChildName() {
+        return childName;
+    }
+
+    public void setChildName(String childName) {
+        this.childName = childName;
+    }
+
+    public Integer getChildScore() {
+        return childScore;
+    }
+
+    public void setChildScore(Integer childScore) {
+        this.childScore = childScore;
+    }
+
+    public String getChildId() {
+        return childId;
+    }
+
+    public void setChildId(String childId) {
+        this.childId = childId;
+    }
+}
