--- app/src/main/java/com/fall2022_group20/dataAccessLayer/Lesson/LessonSchema.java
@@ -1,36 +1,0 @@
-package com.fall2022_group20.dataAccessLayer.Lesson;
-
-import io.realm.RealmList;
-import io.realm.RealmObject;
-import io.realm.annotations.PrimaryKey;
-
-public class LessonSchema extends RealmObject{
-    @PrimaryKey
-    private String lessonId;
-    private Integer points;
-    private String lessonName;
-
-    public String getLessonId() {
-        return lessonId;
-    }
-
-    public void setLessonId(String lessonId) {
-        this.lessonId = lessonId;
-    }
-
-    public Integer getPoints() {
-        return points;
-    }
-
-    public void setPoints(Integer points) {
-        this.points = points;
-    }
-
-    public String getLessonName() {
-        return lessonName;
-    }
-
-    public void setLessonName(String lessonName) {
-        this.lessonName = lessonName;
-    }
-}
