+++ app/src/main/res/values-night/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.Fall2022Group20" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.GrowthPlus" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/dark_green</item>
         <item name="colorPrimaryVariant">@color/light_green</item>
