+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Fall2022-Group20</string>
+    <string name="app_name">GrowthPlus</string>
 </resources>
