+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Fall2022-Group20"
+rootProject.name = "GrowthPlus"
 include ':app'
